Added a key block generator for safecast
This generates a 2048-bit RSA private key plus an ASCII-armored
public key, and a GUID. The format of the block is:

0x0000: RSA private key (in RFC-2440 binary format)
0x0800: Public key in base-64 format (a convenience to send to server)
0x0FF0: 16-byte GUID

The intention is that a device can register itself with the safecast
servers by uploading its public key + GUID. Once it is registered,
all data uploaded by the device can be signed by the device to prove
that the data has not be tampered inside the database or during the
upload process.

Of course, a user can go ahead and replace the GUID or the private
key with their own, and then tamper the data. But the idea is that
the factory records the GUID and public key pairings at manufacturing
time, so such replacement can be identified and potentially eliminated
from the dataset.

So, the purpose of this mechanism isn't security per se, its
tamper-evidence, and traceability to a single device.