Added in crypto code to the mainline branch
Several new commands added in serialinterface.cpp:

TESTSIGN  - performs a test signature operation to exercise crypto blocks
PUBKEY - prints ASCII-armored public key
GUID - prints the device GUID (16 bytes / 128 bits)
KEYVALID - checks if there is a valid private key, or the testing private key, burned into the memory
LOGSIG - computes a signature on the logs

also, placeholder test function stubs for log_read functions have
been added per conversation with Nava.

Note that it's important to add a check to LOGSIG to fail if
the private key is the testing one. Not in there yet because we're
still in testing.