changed genkeys to use absolute paths
Absolute paths are encoded because genkeys is called via udev now,
and we don't do a "proper" installation of genkeys into system path
variables at the moment. And I think I never want to, since this
is really a script to be called on just one machine in the world.