added autotester for automatic testing
This program will interact with the device over the serial port
and issue commands and listen to them.

Currently, it just reprints the test result but string matching
will be added to check for test validity (actually, this is probably
going to be done at the shell script level).

One issue is that devices that hang will hang the test. A timeout
of some sort needs to be implemented to make the test robust.