Add RTC setting to the automatic device burner
Now the RTC will be set to a near-correct unix time at the last
step during burning.

As a side note, ntp should be running on the device doing the
programming to keep the local RTC sync'd with the atomic clocks.
The RTC in most laptops and desktops have abominable drift without
network synchronization!