12.26-b: implement a reset-to-bootloader command on the Onyx.
This makes it possible to program the onyx without any dependency on the
FTDI libs on the computer.

Not thoroughly tested yet, there are a lot of fine points in the reset vector
of the device to check how many times we are allowed to reset before giving up,
do not use this version in production.